``` Refactor booking widget and improve error handling
- Extracted `BookingWidget` component for reuse and modularity.
- Improved error handling in `analytics-0.1.js` with a fallback for Botd loading.
- Added new icons (`link-01`, `link-external-01`) to the icon set.
- Updated `TeamScheduling` to include booking preview and embed options.
- Simplified `BookingPage` by replacing inline logic with `BookingWidget`.
```
