+++ src/routes/src/components/BookingForm/index.ts
@@ -1,1 +1,2 @@
 export * from './BookingForm';
+export * from './BookingWidget';
